refactor: [user]/create-group 적용
