+++ apps/client/src/app/[user]/@modal/(.)create-group/page.tsx
@@ -1,0 +1,3 @@
+import CreateGroupPage from "../../create-group/page";
+
+export default CreateGroupPage;
